backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
flask-cors==4.0.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
zipp==3.21.0
PyJWT==2.8.0
mysql-connector-python==9.4.0

